added namespace api key to nodejs and python runtime packages and added param checks

+++ runtimes/nodejs/actions/object-delete.js
@@ -16,6 +16,12 @@ async function main(args) {
   const { cos, params } = getParamsCOS(args, CloudObjectStorage);
   let response;
   const result = params;
+
+  if (!params.bucket || !params.key || !cos) {
+    result.message = "bucket name, key, and apikey are required for this operation."
+    return result
+  }
+
   try {
     response = await cos.deleteObject({ Bucket: params.bucket, Key: params.key }).promise();
   } catch (err) {
@@ -36,15 +42,33 @@ async function main(args) {
 
 
 function getParamsCOS(args, COS) {
+  var bxCredsApiKey = ""
+  var bxCredsResourceInstanceId = ""
+
+  if (args.__bx_creds && args.__bx_creds['cloud-object-storage']) {
+    if (args.__bx_creds['cloud-object-storage'].apiKey) {
+      bxCredsApiKey = args.__bx_creds['cloud-object-storage'].apikey
+    }
+    if (args.__bx_creds['cloud-object-storage'].resource_instance_id) {
+      bxCredsResourceInstanceId = args.__bx_creds['cloud-object-storage'].resource_instance_id
+    }
+  }
+
   const { bucket, key } = args;
-  const endpoint = args.endpoint || 's3-api.us-geo.objectstorage.softlayer.net';
+  const endpoint = args.endpoint || 's3.us.cloud-object-storage.appdomain.cloud';
   const ibmAuthEndpoint = args.ibmAuthEndpoint || 'https://iam.cloud.ibm.com/identity/token';
-  const apiKeyId = args.apikey || args.apiKeyId || args.__bx_creds['cloud-object-storage'].apikey;
-  const serviceInstanceId = args.resource_instance_id || args.serviceInstanceId || args.__bx_creds['cloud-object-storage'].resource_instance_id;
+  const apiKeyId = args.apikey || args.apiKeyId || bxCredsApiKey || process.env.__OW_IAM_NAMESPACE_API_KEY;
+  const serviceInstanceId = args.resource_instance_id || args.serviceInstanceId || bxCredsResourceInstanceId;
 
   const params = {};
   params.bucket = bucket;
   params.key = key;
+
+  if (!apiKeyId) {
+    const cos = null
+    return { cos, params }
+  }
+
   const cos = new COS.S3({
     endpoint, ibmAuthEndpoint, apiKeyId, serviceInstanceId,
   });
